add leafView / levelView methods again (for compatibility)
The maintainers for the 2.3 release of the Dune core modules simply
refuse to merge the correct renaming from leafView / levelView to
leafGridView / levelGridView, which simply forwards the old methods to
the new ones (on the implementation). Instead, they require the grid
views to have a constructor with fixed arguments.

This patch readds the methods leafView / levelView (forwarding the to
the new methods) on SPGrid again, overriding the wrong default
implementation).

Big thanks for the additional boiler-plate code!