for variable size communication, initialize receiving on wait
Probing for a message causes a non-negligible overhead. As long as the
message fits the eager limit of the MPI implementation, it is much more
efficient to start receiving upon waiting. For larger messages, we still
use Irecv to allow the receive operations to overlap.

For fixed size messages, we still start receiving when communication is
initiated as the size is already known. This way, we circumvent any
problems with the eager limit as long as the user knows the size for
each entity a priori (fixedSize).